Implemented new matrix exponential method: Arnoldi
The new method is based on Arnoldi iteration and doesn't seem to work yet. The problem is, that during construction of the Krylov subspace new elements of the upper Hessenberg matrix get bigger and bigger, diverging to infinity very fast.

Apart from that some arrays throughout the script, which are constructed with np.linspace, had faulty arguments and are now corrected.