Implement new matrix exponential method: Lanczos
The new method is based on Lanczos algorithm and doesn't seem to work yet. The problem is, that convergence cannot be achieved.

Apart from that the Arnoldi method is updated. Even though it doesn't produce errors anymore, the Taylor method leads to completely different results.

The example code now compares Taylor, Arnoldi and Lanczos method with regards to computing time per propagation step and result of propagation.