Implement new reference data for OCS alignment test
As the initially implemented reference data of the chronological sequence of occupation probabilities could not be reproduced by various different versions of RichMol, it is replaced with new reproducible reference data.

Apart from that the test directory '/tests/' as well as implemented tests for the 'tdtools' feature are cleaned up.